added dynamic change trip info
