+++ package.json
@@ -24,6 +24,8 @@
     "node": "10"
   },
   "dependencies": {
+    "chart.js": "2.8.0",
+    "chartjs-plugin-datalabels": "0.7.0",
     "flatpickr": "4.6.2",
     "moment": "2.24.0"
   }
