+++ src/controllers/stats.js
@@ -1,0 +1,2 @@
+import Chart from "chart.js";
+import ChartPlugin from "chartjs-plugin-datalabels";
