+++ src/controllers/trip.js
@@ -1,63 +1,26 @@
-import {cities} from "../site-data";
 import {Position, render, unrender} from "../utils";
 import DaysList from "../components/days-list";
-import Day from "../components/day";
-import PointController from "./point";
-import NewPointController from "./new-point";
+import PointListController from "./point-list";
 import Sort from "../components/sort.js";
+import TripInfo from "../components/trip-info";
 import moment from "moment";
 
 export default class {
-  constructor(container, points, setTotalCost) {
+  constructor(container, points, tripTotalCostElement, tripInfoContainerElement) {
     this._container = container;
     this._points = points.slice(0).sort((a, b) => a.dates.start - b.dates.start);
+    this._pointsSortedByEndDate = this._points.slice(0).sort((a, b) => b.dates.end - a.dates.end);
     this._daysListComponent = new DaysList();
     this._sortComponent = new Sort();
+    this._tripInfoComponent = new TripInfo();
+    this._pointListController = new PointListController(this._daysListComponent.getElement(), this._points, this._onDataChange.bind(this));
 
-    this._subscriptions = [];
-    this._onChangeView = this._onChangeView.bind(this);
-    this._onDataChange = this._onDataChange.bind(this);
-    this._setTotalCost = setTotalCost;
+    this._tripTotalCostElement = tripTotalCostElement;
+    this._tripInfoContainerElement = tripInfoContainerElement;
+    this._tripInfoTitleElement = this._tripInfoComponent.getElement().querySelector(`.trip-info__title`);
+    this._tripInfoDatesElement = this._tripInfoComponent.getElement().querySelector(`.trip-info__dates`);
 
-    this._calculateDurationPoints();
-    this._getUniqueDays();
-    this._getPointsDays();
-  }
-
-  init() {
-    this._uniqueDays.forEach((data, count) => this._createDay(this._pointsDays[count], data, count));
-    this._setTotalCost(this._points);
-
-    render(this._container, this._sortComponent.getElement(), Position.BEFOREEND);
-    render(this._container, this._daysListComponent.getElement(), Position.BEFOREEND);
-
-    this._sortComponent.getElement()
-      .addEventListener(`click`, (evt) => this._onSortInputClick(evt));
-  }
-
-  createNewPoint() {
-    const defaultPoint = {
-      type: {
-        transfer: `Taxi`
-      },
-      city: cities[0],
-      dates: {
-        start: Date.parse(moment()),
-        end: Date.parse(moment().add(1, `hours`)),
-      },
-      photos: new Set(),
-      price: +0,
-      description: ``,
-      options: []
-    };
-
-    this._createPoint(this._daysListComponent.getElement(), defaultPoint, NewPointController);
-  }
-
-  _calculateDurationPoints() {
-    for (const point of this._points) {
-      point.duration = moment(point.dates.end).diff(moment(point.dates.start));
-    }
+    this._init();
   }
 
   show() {
@@ -68,44 +31,48 @@ export default class {
     this._container.classList.add(`trip-events--hidden`);
   }
 
-  _getUniqueDays() {
-    let dates = new Set();
-    this._points.forEach((point) => dates.add(moment(point.dates.start).format(`MMM DD YYYY`)));
-    this._uniqueDays = Array.from(dates);
+  createNewPoint() {
+    this._pointListController.createNewPoint();
   }
 
-  _getPointsDays() {
-    this._pointsDays = this._uniqueDays.map((day) => this._points.filter((point) => moment(point.dates.start).format(`MMM DD YYYY`) === day));
-  }
+  _init() {
+    this._tripTotalCostElement.textContent = this._calculateTotalCost();
+    this._tripInfoTitleElement.textContent = this._getTripInfoTitle();
+    this._tripInfoDatesElement.textContent = this._getTripInfoDates();
 
-  _createDay(points, date, dayNumber) {
-    const day = new Day(points.length, date, dayNumber);
-    const pointsContainers = day.getElement().querySelectorAll(`.trip-events__item`);
-    points.forEach((point, index) => this._createPoint(pointsContainers[index], point, PointController));
-    render(this._daysListComponent.getElement(), day.getElement(), Position.BEFOREEND);
-  }
-
-  _createPoint(container, point, Controller) {
-    const pointController = new Controller(container, point, this._onDataChange, this._onChangeView);
-    this._subscriptions.push(pointController.setDefaultView.bind(pointController));
-  }
-
-  _renderTrip(element) {
-    unrender(this._daysListComponent.getElement());
-    this._daysListComponent.removeElement();
-
-    switch (element.dataset.sortType) {
-      case `event`:
-        this._uniqueDays.forEach((data, count) => this._createDay(this._pointsDays[count], data, count));
-        break;
-      case `price`:
-        this._createDay(this._points.slice(0).sort((a, b) => b.price - a.price));
-        break;
-      case `time`:
-        this._createDay(this._points.sort((a, b) => b.duration - a.duration));
-        break;
-    }
+    render(this._tripInfoContainerElement, this._tripInfoComponent.getElement(), Position.AFTERBEGIN);
+    render(this._container, this._sortComponent.getElement(), Position.BEFOREEND);
     render(this._container, this._daysListComponent.getElement(), Position.BEFOREEND);
+
+    this._sortComponent.getElement()
+      .addEventListener(`click`, (evt) => this._onSortInputClick(evt));
+  }
+
+  _getUniqueCities() {
+    const uniqueCities = new Set();
+    this._points.forEach((point) => uniqueCities.add(point.city));
+    return Array.from(uniqueCities);
+  }
+
+  _getTripInfoTitle() {
+    const uniqueCities = this._getUniqueCities();
+    return uniqueCities.length === 3 ? `${uniqueCities.map((city) => city).join(` — `)}`
+      : `${this._points[0].city} — ... — ${this._pointsSortedByEndDate[0].city}`;
+  }
+
+  _getTripInfoDates() {
+    return `${moment(this._points[0].dates.start).format(`MMM DD`)} — ${moment(this._pointsSortedByEndDate[0].dates.end).format(`DD MMM`)}`;
+  }
+
+  _calculateTotalCost() {
+    let cost = 0;
+    for (let point of this._points) {
+      cost += point.price;
+      point.options.filter((option) => option.isApplied).forEach((option) => {
+        cost += option.price;
+      });
+    }
+    return cost;
   }
 
   _onDataChange(newData, oldData) {
@@ -117,18 +84,17 @@ export default class {
     } else {
       this._points[index] = newData;
     }
-
     this._points.sort((a, b) => a.dates.start - b.dates.start);
-    this._calculateDurationPoints();
-    this._getUniqueDays();
-    this._getPointsDays(this._uniqueDays);
-    this._setTotalCost(this._points);
+    this._tripTotalCostElement.textContent = this._calculateTotalCost();
 
-    this._renderTrip(this._sortComponent.getElement().querySelector(`.trip-sort__input:checked`));
-  }
+    this._pointsSortedByEndDate = this._points.slice(0).sort((a, b) => b.dates.end - a.dates.end);
+    this._tripInfoTitleElement.textContent = this._getTripInfoTitle();
+    this._tripInfoDatesElement.textContent = this._getTripInfoDates();
 
-  _onChangeView() {
-    this._subscriptions.forEach((it) => it());
+    unrender(this._daysListComponent.getElement());
+    this._daysListComponent.removeElement();
+    this._pointListController.renderPointList(this._sortComponent.getElement().querySelector(`.trip-sort__input:checked`), this._points, this._daysListComponent.getElement());
+    render(this._container, this._daysListComponent.getElement(), Position.BEFOREEND);
   }
 
   _onSortInputClick(evt) {
@@ -137,6 +103,9 @@ export default class {
       return;
     }
 
-    this._renderTrip(evt.target);
+    unrender(this._daysListComponent.getElement());
+    this._daysListComponent.removeElement();
+    this._pointListController.renderPointList(evt.target, this._points, this._daysListComponent.getElement());
+    render(this._container, this._daysListComponent.getElement(), Position.BEFOREEND);
   }
 }
